fix: move split panel behavior tests into panel-behavior subdirectory
The test split pushed factory-session-detail/components to 12 direct files,
tripping the 10-file feature-folder file-count guard in ui-lint.
